Init Injectory Dependecy

+++ Cap01/MiPrimeraWeb/Cibertec.Model/OrderDetails.cs
@@ -9,8 +9,8 @@ namespace Cibertec.Models
         public int Id { get; set; }
         public int OrderId { get; set; }
         public int ProductId { get; set; }
-        public int UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Quantity { get; set; }
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
     } 
 }
